Fixing bugs on V1
