--- JS/UserInput.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-
-
-
-<form>
-	<p style = "color:white"> Username : </p> <br>
-	<input type = "text"  id = "userName">  <br> 
-	<p style = "color:white"> Password : </p> <br> 
-	<input type = "Password" id = "Password"> <br>
-	<p style = "color:white">Norad ID: </p> <br>
-	<input type="text" id="NoradId"> <br>
-</form>
-
-<br>
-
-<button style = "color:red" onclick="myFunction()">Track</button>
-
-<p id="demo"></p>
-
-
-
-
-<script>
-function myFunction() {
-
-  
-  var x = document.getElementById("NoradId").value;
-  console.log(x);
-  update(x);
-  ButtonHandler(document.getElementById("userName").value, document.getElementById("Password").value , document.getElementById("NoradId").value)
-  document.getElementById("userName").value ='';
-  document.getElementById("Password").value = '';
-  document.getElementById("NoradId").value = '' ;
-  
-}
-
-longLat = {long:0,lat:0};
-
-
-</script>
-
-<script src = "app1.js"> </script>
-<script src="node_modules/openmct/dist/openmct.js"></script>
-<script src="node_modules/satellite.js/dist/satellite.min.js"></script>
-<script src="tleToCoordinates.js"></script>
-<script src = "ButtonHandler.js"> </script>
-<script src="./webMap/res/webMap.js"></script>
-
-
-
-</body>
-</html>
