+++ JS/V1/trial.html
@@ -9,11 +9,12 @@
 
     <div id="map"></div>
 
-    <script src="./webMap/res/webMap.js"></script>
+ 
 
     <script async defer
     src="https://maps.googleapis.com/maps/api/js?key=AIzaSyD2QUO09KrtwuLJEdTx6p-PpGjYUhbaA5o&callback=initMap">
     </script>
 
+    <script src="./webMap/res/webMap.js"></script>
   </body>
 </html>
