+++ JS/V1/webMap/res/webMap.js
@@ -1,14 +1,18 @@
 //global val to be updated by satellite.js output
-let satLoc = {lat: -10.97, lng: -1.28};
+var satLoc = {lat: -10.97, lng: -1.28};
+
+var marker = new google.maps.Marker({});
 
 function update(satLocation) {
 	satLoc = satLocation;
-	console.log(satLoc);
+  console.log(satLoc);
+	console.log(satLocation);
+  marker.setPosition(satLoc);
 }
 //initialises the map
 function initMap() {
 var map = new google.maps.Map(document.getElementById('map'), {
-  zoom: 2,
+  zoom: 1,
   center: satLoc,
    mapTypeId:'satellite',
    mapTypeControl: true,
@@ -19,20 +23,25 @@ var map = new google.maps.Map(document.getElementById('map'), {
    streetViewControl: false
  });
 
-var ship = './webMap/res/ship.png';
-drawLatLng(satLoc, ship, map);
+  var icon = './webMap/res/ship.png';
+  marker = new google.maps.Marker({
+    position:satLoc,
+    map:map,
+    icon:icon
+  });
+
+    
+
+    // drawLatLng( satLoc,marker, map);
  } //end initMap
 
-//function to draw Marker which refreshes every 1000ms
-function drawLatLng(satLoc,icon,map){
-    window.setInterval(function() {
-      // satLoc['lat'] = satLoc['lat']+1;
+// window.setInterval(function(){
 
-      var marker = new google.maps.Marker(
-       {
-         position: satLoc,
-         map: map,
-         icon:icon,
-       });
-  }, 1000);
-} //end drawLatLng
+//   console.log(window.satLoc);
+// }, 1000);
+
+//function to draw Marker which refreshes every 1000ms
+// function drawLatLng(satLoc,marker,map){
+
+   
+// } //end drawLatLng
