--- JS/index.html
@@ -1,80 +1,0 @@
-
-<!doctype html>
-<html lang="en">
-<head>
-
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
-    <title></title>
-    <script src="node_modules/openmct/dist/openmct.js"></script>
-    <script src="node_modules/satellite.js/dist/satellite.min.js"></script>
-	 <script src="tleToCoordinates.js"></script>
-	
-	
-	
-    <script>
-        var THIRTY_MINUTES = 30 * 60 * 1000;
-        document.addEventListener("DOMContentLoaded", function () {
-            //openmct.install(new OpenMCTMapPlugin());
-			
-            openmct.install(openmct.plugins.MyItems());
-            openmct.install(openmct.plugins.LocalStorage());
-            openmct.install(openmct.plugins.Espresso());
-            openmct.install(openmct.plugins.Generator());
-            openmct.install(openmct.plugins.ExampleImagery());
-            openmct.install(openmct.plugins.UTCTimeSystem());
-            openmct.install(openmct.plugins.ImportExport());
-		
-            openmct.install(openmct.plugins.Conductor({
-                menuOptions: [
-                    {
-                        name: "Fixed",
-                        timeSystem: 'utc',
-                        bounds: {
-                            start: Date.now() - 30 * 60 * 1000,
-                            end: Date.now()
-                        }
-                    },
-                    {
-                        name: "Realtime",
-                        timeSystem: 'utc',
-                        clock: 'local',
-                        clockOffsets: {
-                            start: -25 * 60 * 1000,
-                            end: 5 * 60 * 1000
-                        }
-                    }
-                ]
-            }));
-            openmct.time.clock('local', {start: -THIRTY_MINUTES, end: 0});
-            openmct.time.timeSystem('utc');
-            openmct.setAssetPath("node_modules/openmct/dist");
-            openmct.start();
-
-
-          });
-    </script>
-    <link rel="stylesheet" href="node_modules/openmct/dist/platform/commonUI/general/res/css/startup-base.css">
-    <link rel="stylesheet" href="node_modules/openmct/dist/platform/commonUI/general/res/css/openmct.css">
-    <link rel="icon" type="image/png" href="node_modules/openmct/dist/platform/commonUI/general/res/images/favicons/favicon-32x32.png" sizes="32x32">
-    <link rel="icon" type="image/png" href="node_modules/openmct/dist/platform/commonUI/general/res/images/favicons/favicon-96x96.png" sizes="96x96">
-    <link rel="icon" type="image/png" href="node_modules/openmct/dist/platform/commonUI/general/res/images/favicons/favicon-16x16.png" sizes="16x16">
-    <link rel="shortcut icon" href="node_modules/openmct/dist/platform/commonUI/general/res/images/favicons/favicon.ico">
-</head>
-<body class="user-environ">
-    <div class="l-splash-holder s-splash-holder">
-        <div class="l-splash s-splash"></div>
-    </div>
-</body>
-
-			<input type="text" placeholder="Enter a value">
-<br><br>
-			
-</html>
-
-
-
-
-
-
-
