--- JS/trial.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title></title>
-    <link rel="stylesheet" href="./webMap/res/webMap.css">
-  </head>
-  <body>
-
-    <div id="map"></div>
-
- 
-
-    <script async defer
-    src="https://maps.googleapis.com/maps/api/js?key=AIzaSyD2QUO09KrtwuLJEdTx6p-PpGjYUhbaA5o&callback=initMap">
-    </script>
-
-    <script src="./webMap/res/webMap.js"></script>
-  </body>
-</html>
