--- JS/webMap/res/webMap.css
@@ -1,9 +1,0 @@
-html, body {
-        height: 100%;
-        margin: 0;
-        padding: 0;
-}
-
-#map {
-  height: 100%;
-}
